Fix the type of a statistics counter (unsigned --> unsigned long).

+++ lib/libc/stdlib/malloc.c
@@ -310,7 +310,7 @@ struct malloc_bin_stats_s {
 	 * during normal operation, so is maintained here in order to allow
 	 * calculating the high water mark.
 	 */
-	unsigned	curcached;
+	unsigned long	curcached;
 };
 
 typedef struct arena_stats_s arena_stats_t;
